Lets Encrypt Email disabled

+++ security/firewall-manager.php
@@ -45,7 +45,7 @@ class rsssl_firewall_manager {
 		if ( !get_site_option("rsssl_header_detection_nonce")) {
 			update_site_option("rsssl_header_detection_nonce", rand(1000, 999999999) );
 		}
-		return get_site_option("rsssl_header_detection_nonce");
+		return (int) get_site_option("rsssl_header_detection_nonce");
 	}
 
 	/**
